Struct partiql_logical::LogicalPlan

pub struct LogicalPlan<T>where
    T: Default,{ /* private fields */ }

Represents a PartiQL logical plan.

Implementations

impl<T> LogicalPlan<T>where T: Default,

pub fn new() -> Self

Creates a new default logical plan.

pub fn add_operator(&mut self, op: T) -> OpId

Adds a new operator to the plan.

pub fn add_flow(&mut self, src: OpId, dst: OpId)

Adds a data flow to the plan.

pub fn add_flow_with_branch_num(&mut self, src: OpId, dst: OpId, branch_num: u8)

Adds a data flow with a branch number.

pub fn extend_with_flows(&mut self, flows: &[(OpId, OpId)])

Extends the logical plan with the given data flows. #Examples:

use partiql_logical::{BindingsOp, GroupBy, GroupingStrategy, LimitOffset, LogicalPlan, OrderBy};
let mut p: LogicalPlan<BindingsOp> = LogicalPlan::new();

let a = p.add_operator(BindingsOp::OrderBy(OrderBy{specs: vec![]}));
let b = p.add_operator(BindingsOp::Sink);
let c = p.add_operator(BindingsOp::LimitOffset(LimitOffset{limit:None, offset:None}));
let d = p.add_operator(BindingsOp::GroupBy(GroupBy {
    strategy: GroupingStrategy::GroupFull,
    exprs: Default::default(),
    aggregate_exprs: vec![],
    group_as_alias: None,
}));

p.add_flow(a, b);

p.extend_with_flows(&[(a,c), (c, d)]);
assert_eq!(3, p.flows().len());

pub fn operator_count(&self) -> usize

Returns the number of operators in the plan.

pub fn operators(&self) -> &Vec<T>

Returns the operators of the plan.

pub fn operators_by_id(&self) -> impl Iterator<Item = (OpId, &T)>

Returns the operators of the plan with their OpId.

pub fn flows(&self) -> &Vec<(OpId, OpId, u8)>

Returns the data flows of the plan.

pub fn operator(&self, id: OpId) -> Option<&T>

pub fn operator_as_mut(&mut self, id: OpId) -> Option<&mut T>